requirements.txt
fastapi~=0.111.0
uvicorn
easyocr
pillow~=10.3.0
torchvision~=0.18.1
onnxruntime~=1.17.3
numpy~=1.26.4
requests~=2.32.3
python-dotenv~=1.0.1
torch~=2.3.1
starlette~=0.37.2
gdown~=5.2.0
onnx~=1.16.1